v2 reporter: code splitting into different files
